+++ src/plugins/plot/barGraph/inspector/SeriesOptions.vue
